% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.h.R
\name{Regression}
\alias{Regression}
\title{Regression}
\usage{
Regression(
  data,
  dep,
  covs = NULL,
  factors = NULL,
  stdCovs = list(list()),
  encCat = list(list()),
  models,
  hidden_neurons_tune = "none",
  learning_rate_tune = "none",
  act_funcs,
  fix_n_neurons,
  fix_learning_rate,
  min_n_neurons = 5,
  max_n_neurons = 20,
  min_learning_rate = -3,
  max_learning_rate = -1,
  tuners = "none",
  metrics = "none",
  plot_results = FALSE
)
}
\arguments{
\item{data}{the data as a data frame}

\item{dep}{the dependent variable from \code{data}, variable must be
numeric}

\item{covs}{the covariates from \code{data}}

\item{factors}{the fixed factors from \code{data}}

\item{stdCovs}{.}

\item{encCat}{.}

\item{models}{.}

\item{hidden_neurons_tune}{.}

\item{learning_rate_tune}{.}

\item{act_funcs}{.}

\item{fix_n_neurons}{.}

\item{fix_learning_rate}{.}

\item{min_n_neurons}{a number between 50 and 99.9 (default: 95) specifying
the confidence interval width}

\item{max_n_neurons}{a number between 50 and 99.9 (default: 95) specifying
the confidence interval width}

\item{min_learning_rate}{a number between 50 and 99.9 (default: 95)
specifying the confidence interval width}

\item{max_learning_rate}{a number between 50 and 99.9 (default: 95)
specifying the confidence interval width}

\item{tuners}{.}

\item{metrics}{.}

\item{plot_results}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Regression
}
